+++ app/services/sales-api/handlers/debug/checkgrp/checkgrp.go
@@ -8,7 +8,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/ardanlabs/service/business/data/database"
+	"github.com/ardanlabs/service/business/sys/database"
 	"github.com/jmoiron/sqlx"
 	"go.uber.org/zap"
 )
