+++ app/tooling/sales-admin/commands/migrate.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/ardanlabs/service/business/data/database"
 	"github.com/ardanlabs/service/business/data/dbschema"
+	"github.com/ardanlabs/service/business/sys/database"
 )
 
 // ErrHelp provides context that help was given.
