+++ app/tooling/sales-admin/commands/useradd.go
@@ -6,8 +6,8 @@ import (
 	"time"
 
 	"github.com/ardanlabs/service/business/core/user"
-	"github.com/ardanlabs/service/business/data/database"
 	"github.com/ardanlabs/service/business/sys/auth"
+	"github.com/ardanlabs/service/business/sys/database"
 	"go.uber.org/zap"
 )
