+++ business/data/dbtest/dbtest.go
@@ -13,9 +13,9 @@ import (
 	"time"
 
 	"github.com/ardanlabs/service/business/core/user/dbuser"
-	"github.com/ardanlabs/service/business/data/database"
 	"github.com/ardanlabs/service/business/data/dbschema"
 	"github.com/ardanlabs/service/business/sys/auth"
+	"github.com/ardanlabs/service/business/sys/database"
 	"github.com/ardanlabs/service/foundation/docker"
 	"github.com/ardanlabs/service/foundation/keystore"
 	"github.com/ardanlabs/service/foundation/logger"
